pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mqtt-house"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [{ name = "Mark Hall", email = "mark.hall@work.room3b.eu" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "fastapi[all]",
  "httpx",
  "typer[all]",
  "pydantic>=2,<3",
  "PyYaml>=6,<7",
  "rshell",
]

[project.urls]
Documentation = "https://github.com/unknown/mqtt-house#readme"
Issues = "https://github.com/unknown/mqtt-house/issues"
Source = "https://github.com/unknown/mqtt-house"

[project.scripts]
mqtt-house = "mqtt_house.cli:app"

[tool.hatch.version]
path = "mqtt_house/__about__.py"

[tool.hatch.envs.default]
extra-dependencies = ["coverage[toml]>=6.5", "pytest", "ruff>=0.5.0"]

[tool.hatch.envs.default.scripts]
mqtt-house = "python -m mqtt_house {args}"
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = ["- coverage combine", "coverage report"]
cov = ["test-cov", "cov-report"]
check-style = ["ruff check {args:.}"]
format-style = ["ruff format {args:.}"]


[tool.ruff]
target-version = "py37"
line-length = 120
lint.select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
lint.ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore string concatentation
  "ISC001",
  # Ignore checks for possible passwords
  "S105",
  "S106",
  "S107",
  # Ignore complexity
  "C901",
  "PLR0911",
  "PLR0912",
  "PLR0913",
  "PLR0915",
]
lint.unfixable = [
  # Don't touch unused imports
  "F401",
]
exclude = ["mqtt_house/micro"]

[tool.ruff.lint.isort]
known-first-party = ["mqtt_house"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.coverage.run]
source_pkgs = ["mqtt_house", "tests"]
branch = true
parallel = true
omit = ["mqtt_house/__about__.py"]

[tool.coverage.paths]
mqtt_house = ["mqtt_house", "*/mqtt-house/mqtt_house"]
tests = ["tests", "*/mqtt-house/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
